restructure packaging xml to input
