+++ src/main/java/input/file/XlsxFile.java
@@ -1,4 +1,4 @@
-package xml.filetype;
+package input.file;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -7,35 +7,35 @@ import java.util.Iterator;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import org.apache.poi.hssf.usermodel.HSSFSheet;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
 import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.xssf.usermodel.XSSFSheet;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
 /**
- * Class to represent a Microsoft Excel (xls) file.
+ * Class to represent a Microsoft excel (xlsx) ooxml file.
  * @author Paul
  *
  */
-public class XlsFile extends ExcelFile {
+public class XlsxFile extends ExcelFile {
 
-	private Logger logger = Logger.getLogger("XlsFile");
+	private Logger logger = Logger.getLogger("XlsxFile");
 	
-	public XlsFile(String path) {
+	public XlsxFile(String path) {
 		super(path);
 	}
 
 	@Override
 	public InputStream getDataStream() throws IOException {
-		
+
 		FileInputStream file = new FileInputStream(getFile());
 		try {
-			workbook = new HSSFWorkbook(file);
-			HSSFSheet sheet = (HSSFSheet) workbook.getSheetAt(0);
+			workbook = new XSSFWorkbook(file);
+			XSSFSheet sheet = (XSSFSheet) workbook.getSheetAt(0);
 			Iterator<Row> rowIterator = sheet.iterator();
 			return createStream(rowIterator);
 			
 		} catch (IOException e) {
-			logger.log(Level.SEVERE, "Error reading xls file", e);
+			logger.log(Level.SEVERE, "Error reading xlsx file", e);
 			throw e;
 		}
 	}
