+++ src/main/java/input/reader/XmlReader.java
@@ -1,11 +1,11 @@
-package xml;
+package input.reader;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
-import xml.filetype.DataFile;
+import input.file.DataFile;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
