+++ src/test/java/input/file/DataFileTest.java
@@ -1,6 +1,10 @@
-package xml.filetype;
+package input.file;
 
 import static org.junit.Assert.*;
+import input.file.DataFile;
+import input.file.PlainTextFile;
+import input.file.XlsFile;
+import input.file.XlsxFile;
 
 import java.io.*;
