+++ src/test/java/input/file/PlainTextFileTest.java
@@ -1,6 +1,7 @@
-package xml.filetype;
+package input.file;
 
 import static org.junit.Assert.*;
+import input.file.PlainTextFile;
 
 import java.io.BufferedReader;
 import java.io.IOException;
