+++ src/test/java/input/reader/XmlReaderTest.java
@@ -1,8 +1,10 @@
-package xml;
+package input.reader;
+
+import input.file.DataFile;
+import input.reader.XmlReader;
 
 import java.io.File;
 import java.io.IOException;
-
 import java.util.List;
 
 import javax.xml.parsers.ParserConfigurationException;
@@ -14,7 +16,6 @@ import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
-import xml.filetype.DataFile;
 import static org.junit.Assert.*;
 
 /**
